requirements.txt
numpy
pandas
torch
transformers
huggingface_hub
accelerate
safetensors
bitsandbytes            
sentence_transformers
optimum
onnxruntime
langchain
langchain_core
langchain_community
langchain_chroma        
langchain_huggingface
langsmith               
chromadb
rank-bm25
fastapi
uvicorn
pydantic
onnx
python-dotenv
openai
langchain-openai